+++ test/write_tests.h
@@ -1,2 +1,4 @@
-void write_test(int repeats, int start, int last, int inc);
-void write_blocksize_test(int repeats, int size, int start, int last, int inc);
+#include <stddef.h>
+
+void write_test(int repeats, size_t start, size_t last, size_t inc);
+void write_blocksize_test(int repeats, size_t size, size_t start, size_t last, size_t inc);
